+++ config.yaml
@@ -1,15 +1,40 @@
 Arguments:
-  filename: image.png
-  contrast: 1
-  sampleSize: 16x16
-  contrastbreak: 128
+  filename: 
+    default: image.png
+    help: "Filepath to your image, the default image name is image.png"
   output: 
     default: null
     const: output.txt
-  blur: 30
-  palettename: null
-  hide: False
-  characters: \'^:!=*$%@#
-  characterfile: null
-  foreground: True
-  background: True
+    help: "Specify output file it can be then displayed with `cat output.txt` with all the colors"
+  
+  contrast:
+    default: 1
+    help: "allows you to change the contrast of the image for better results. (recomended 1 - 1.2 range)"
+  sampleSize: 
+    default: 16x16
+    help: "Size of the samples, default is 16x16 (the output will be 16x smaller) enter as XxY or just X"
+  contrastbreak: 
+    default: 128
+    help: "Border of darkness levels between making a pixel darker or brighter (0-255 recomended range 50-200)"
+  blur: 
+    default: 30
+    help: "Blurs a furhter range of colors together"
+
+  hide: 
+    default: False
+    help: "if false (default) will display the image as it is being generated"
+  palettename: 
+    default: null
+    help: "Enter name of the pallete from palettes or file path"
+  characters: 
+    default: \'^:!=*$%@#
+    help: "enter characters"
+  characterfile:
+    default: null
+    help: "enter filepath/name with the characters"
+  foreground: 
+    default: True
+    help: "Doesn't display any foreground colors/symbols"
+  background: 
+    default: True
+    help: "Doesn't display any background colors, bg colors will transition into foreground"
